moving around file to create eeprom module
