+++ src/edgepi/eeprom/eeprom_constants.py
