+++ src/edgepi/eeprom/eeprom_mapping.proto
