+++ src/edgepi/eeprom/protobuf_mapping.py
@@ -3,7 +3,7 @@
 # pylint: disable=no-name-in-module
 # https://github.com/protocolbuffers/protobuf/issues/10372
 from dataclasses import dataclass
-from edgepi.calibration.eeprom_mapping_pb2 import EepromLayout
+from edgepi.eeprom.eeprom_mapping_pb2 import EepromLayout
 from edgepi.calibration.calibration_constants import CalibParam
 
 @dataclass
