+++ src/test_edgepi/unit_tests/test_dac/test_edgepi_dac.py
@@ -25,9 +25,9 @@ from edgepi.dac.dac_constants import (
 )
 from edgepi.dac.edgepi_dac import EdgePiDAC
 from edgepi.calibration.calibration_constants import CalibParam
-from edgepi.calibration.protobuf_mapping import EdgePiEEPROMData
-from edgepi.calibration.eeprom_mapping_pb2 import EepromLayout
-from test_edgepi.unit_tests.test_calibration.read_serialized import read_binfile
+from edgepi.eeprom.protobuf_mapping import EdgePiEEPROMData
+from edgepi.eeprom.eeprom_mapping_pb2 import EepromLayout
+from test_edgepi.unit_tests.test_eeprom.read_serialized import read_binfile
 
 dummy_calib_param_dict = {0:CalibParam(gain=1,offset=0),
                           1:CalibParam(gain=1,offset=0),
