+++ src/test_edgepi/unit_tests/test_eeprom/read_serialized.py
