+++ tests/hardware_tests/test_eeprom_userspace.py
@@ -13,8 +13,8 @@ import os
 PATH = os.path.dirname(os.path.abspath(__file__))
 
 import pytest
-from edgepi.calibration.eeprom_constants import EdgePiMemoryInfo, EEPROMInfo
-from edgepi.calibration.edgepi_eeprom import EdgePiEEPROM
+from edgepi.eeprom.eeprom_constants import EdgePiMemoryInfo, EEPROMInfo
+from edgepi.eeprom.edgepi_eeprom import EdgePiEEPROM
 
 
 @pytest.fixture(name="eeprom")
